Pull testing mini-funk into separate file for reuse in val unit testing
